requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-uuslug==1.2.0
idna==2.10
importlib-metadata==2.0.0
Markdown==3.2.2
mysqlclient==2.0.1
Pillow==7.2.0
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
zipp==3.2.0
